added session token field to user db, added functionality to make new sessiontokens on log in\
